Remove deferred Teaching sections from navigation
